requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.0
cryptography==44.0.1
fastapi==0.115.12
ffmpeg==1.4
filelock==3.17.0
fsspec==2025.2.0
h11==0.16.0
idna==3.10
Jinja2==3.1.5
langdetect==1.0.9
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.4.2
numpy==2.2.3
psycopg2-binary
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyserial==3.5
python-multipart==0.0.20
requests==2.32.3
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
SQLAlchemy==2.0.38
srt==3.5.3
starlette==0.46.2
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.2
vosk==0.3.44
websockets==15.0
